services/backend-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
loguru==0.7.0
python-multipart==0.0.6
strands-agents[ollama]
strands-agents-tools
ollama>=0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
html2text==2020.1.16
